# docs/swagger.yaml
definitions:
  model.JWT:
    properties:
      access_token:
        type: string
      id:
        type: string
      refresh_token:
        type: string
    required:
    - access_token
    - id
    - refresh_token
    type: object
  model.Profile:
    properties:
      avatar:
        type: string
      deaths:
        type: integer
      headshot_rate:
        type: integer
      id:
        type: string
      kills:
        type: integer
      name:
        type: string
      url:
        type: string
    required:
    - avatar
    - deaths
    - headshot_rate
    - id
    - kills
    - name
    - url
    type: object
  render.Err:
    properties:
      code:
        type: integer
      message:
        type: string
    required:
    - code
    - message
    type: object
info:
  contact: {}
  title: Backend API
  version: "1.0"
paths:
  /api/auth/login:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "302":
          description: Found
        "405":
          description: Method Not Allowed
          schema:
            $ref: '#/definitions/render.Err'
      summary: Redirects client to Steam authentication page
      tags:
      - auth
  /api/auth/process:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.JWT'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/render.Err'
        "405":
          description: Method Not Allowed
          schema:
            $ref: '#/definitions/render.Err'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/render.Err'
      summary: Processes Steam authentication response and generates JWT tokens
      tags:
      - auth
  /api/auth/refresh:
    post:
      parameters:
      - description: User ID
        in: formData
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.JWT'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/render.Err'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/render.Err'
        "405":
          description: Method Not Allowed
          schema:
            $ref: '#/definitions/render.Err'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/render.Err'
      security:
      - BearerAuth: []
      summary: Refreshes JWT tokens
      tags:
      - auth
  /api/profile/{id}:
    get:
      consumes:
      - application/json
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Profile'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/render.Err'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/render.Err'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/render.Err'
      security:
      - BearerAuth: []
      summary: Retrieves user profile
      tags:
      - profile
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"